requirements.txt
pbr!=2.1.0,>=2.0.0 # Apache-2.0
SQLAlchemy>=1.2.19 # MIT
keystonemiddleware>=4.20.0 # Apache-2.0
greenlet>=0.4.10,!=0.4.14 # MIT
keystoneauth1>=3.16.0 # Apache-2.0
oslo.config>=6.1.0 # Apache-2.0
oslo.context>=2.21.0 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.serialization!=2.19.1,>=2.21.1 # Apache-2.0
oslo.utils>=3.40.2 # Apache-2.0
oslo.db>=4.44.0 # Apache-2.0
oslo.messaging>=10.3.0 # Apache-2.0
oslo.policy>=2.3.0 # Apache-2.0
oslo.i18n>=3.15.3 # Apache-2.0
oslo.service>=1.40.1 # Apache-2.0
oslo.middleware>=3.31.0 # Apache-2.0
oslo.versionedobjects>=1.35.0 # Apache-2.0
openstacksdk>=0.35.0 # Apache-2.0
